Save state in redis when updatingDockers

+++ crons/updateDockers.ts
@@ -9,6 +9,7 @@ import { maxPromises } from "duplicatedFilesCleanerIncognito";
 import { dataDir } from "../incognito/docker/createDocker.ts";
 import { Info, docker } from "duplicatedFilesCleanerIncognito";
 import duplicatedFilesCleaner from "../duplicatedFilesCleaner.ts";
+import { UpdatingNode, updatingNodes } from "../utils/variables.ts";
 import { changeNode, getNodes } from "../controllers/node.controller.ts";
 import createDockerAndConfigs, { addNodeToConfigs } from "../incognito/createDockerAndConfigs.ts";
 import deleteDockerAndConfigs, { removeNodeFromConfigs } from "../incognito/deleteDockerAndConfigs.ts";
@@ -27,8 +28,16 @@ async function createTempDataDir(startText = "") {
   return { fullPath, tempDirName };
 }
 
+async function updateTagInDB(node: Node) {
+  const { [node.dockerIndex]: nodeInfo } = await duplicatedFilesCleaner.getInfo([node.dockerIndex]);
+  await changeNode({ _id: node._id }, { $set: { dockerTag: nodeInfo.docker.tag } });
+}
+
+// restore previous state if any
+setTimeout(() => updateDockers({ force: true, dockerIndexes: updatingNodes.map((n) => n.dockerIndex) }), 3_000);
+
 export default async function updateDockers({ force = false, dockerIndexes }: UpdateDockersOptions = {}) {
-  if (instanceRunning) return;
+  if (instanceRunning || dockerIndexes?.length === 0) return;
   instanceRunning = true;
 
   const latestTag = await getLatestTag();
@@ -55,35 +64,56 @@ export default async function updateDockers({ force = false, dockerIndexes }: Up
 
         console.log(`Updating node ${node.dockerIndex} from ${info.docker.tag} to ${latestTag}`);
 
+        const updatingInfo: UpdatingNode = updatingNodes.find((n) => n.dockerIndex === dockerIndex) || {
+          dockerIndex,
+          tempDir: null,
+          deleted: false,
+          created: false,
+          backupRestored: false,
+          provenToBeOnline: false,
+          dockerWasOnline: info.docker.running,
+        };
+
         // save the beacon and shard files somewhere else
-        const tempDir = await createTempDataDir(`node_${node.dockerIndex}`).then((dir) => dir.fullPath);
-        const blockDir = `${dataDir}_${node.dockerIndex}/mainnet/block`;
-        const tempBlockDir = join(tempDir, "block");
-        const backup = await doesDirExists(blockDir);
-        if (backup) await Deno.rename(blockDir, tempBlockDir);
+        const { tempDir, blockDir, tempBlockDir, backup } = await (async () => {
+          if (updatingInfo.tempDir) return updatingInfo.tempDir;
+
+          const tempDir = await createTempDataDir(`node_${node.dockerIndex}`).then((dir) => dir.fullPath);
+          const blockDir = `${dataDir}_${node.dockerIndex}/mainnet/block`;
+          const tempBlockDir = join(tempDir, "block");
+          const backup = await doesDirExists(blockDir);
+          if (backup) await Deno.rename(blockDir, tempBlockDir);
+
+          return (updatingInfo.tempDir = { tempDir, blockDir, tempBlockDir, backup });
+        })();
 
         // delete and recreate the docker
-        removeNodeFromConfigs(node.dockerIndex);
-        await deleteDockerAndConfigs({
-          number: node.number,
-          clientId: node.client,
-          dockerIndex: node.dockerIndex,
-        });
-        await createDockerAndConfigs({
-          inactive: true,
-          nodeId: node._id,
-          number: node.number,
-          clientId: node.client,
-          rcpPort: node.rcpPort,
-          validator: node.validator,
-          dockerIndex: node.dockerIndex,
-          validatorPublic: node.validatorPublic,
-        });
-        // await changeNode({ _id: node._id }, { $set: { inactive: true } });
-        // await deleteDocker(node.dockerIndex);
-        // await createDocker(node.rcpPort, node.validatorPublic, node.dockerIndex, latestTag);
+        if (!updatingInfo.deleted) {
+          removeNodeFromConfigs(node.dockerIndex);
+          await deleteDockerAndConfigs({
+            number: node.number,
+            clientId: node.client,
+            dockerIndex: node.dockerIndex,
+          });
+          updatingInfo.deleted = true;
+        }
+        if (!updatingInfo.created) {
+          await createDockerAndConfigs({
+            inactive: true,
+            nodeId: node._id,
+            number: node.number,
+            clientId: node.client,
+            rcpPort: node.rcpPort,
+            validator: node.validator,
+            dockerIndex: node.dockerIndex,
+            validatorPublic: node.validatorPublic,
+          });
+          updatingInfo.created = true;
+        }
 
-        await restoreAndExit(node, backup, blockDir, tempBlockDir, tempDir, info).catch(handleError);
+        await restoreAndExit(node, backup, blockDir, tempBlockDir, tempDir, info, updatingInfo, latestTag).catch(
+          handleError
+        );
       } catch (e) {
         await handleError(e);
       }
@@ -94,48 +124,48 @@ export default async function updateDockers({ force = false, dockerIndexes }: Up
   instanceRunning = false;
 }
 
-async function updateTagInDB(node: Node) {
-  const { [node.dockerIndex]: nodeInfo } = await duplicatedFilesCleaner.getInfo([node.dockerIndex]);
-  await changeNode({ _id: node._id }, { $set: { dockerTag: nodeInfo.docker.tag } });
-}
-
 async function restoreAndExit(
   node: Node,
   backup: boolean,
   blockDir: string,
   tempBlockDir: string,
   tempDir: string,
-  info: Info
+  info: Info,
+  updatingInfo: UpdatingNode,
+  latestTag: string
 ) {
   const { dockerIndex } = node;
 
   // wait for it to be online in the monitor
-  await (async (timedout = false) => {
-    await Promise.race([
-      (async () => {
-        while (!timedout) {
-          const { [dockerIndex]: newInfo } = await duplicatedFilesCleaner.getInfo([dockerIndex]);
-          if (!newInfo?.docker.running) {
-            if (await doesDirExists(`${dataDir}_${dockerIndex}`))
-              await Deno.remove(`${dataDir}_${dockerIndex}`, { recursive: true });
-            await docker(`inc_mainnet_${dockerIndex}`, "start");
-          }
+  if (!updatingInfo.provenToBeOnline) {
+    await (async (timedout = false) => {
+      await Promise.race([
+        (async () => {
+          while (!timedout) {
+            const { [dockerIndex]: newInfo } = await duplicatedFilesCleaner.getInfo([dockerIndex]);
+            if (!newInfo?.docker.running) {
+              if (await doesDirExists(`${dataDir}_${dockerIndex}`))
+                await Deno.remove(`${dataDir}_${dockerIndex}`, { recursive: true });
+              await docker(`inc_mainnet_${dockerIndex}`, "start");
+            }
 
-          const [nodeStatus] = await getNodesStatus({ dockerIndexes: [dockerIndex], onlyActive: false });
-          if (!nodeStatus) console.error(new Error(`Node ${dockerIndex} not found in the monitor`));
-          if (nodeStatus.status === "ONLINE") break;
-          await sleep(5);
-        }
-      })(),
-      sleep(60 * 5).finally(() => {
-        timedout = true;
-        console.error(new Error(`Node ${dockerIndex} did not come online in 5 minutes.`));
-      }),
-    ]);
-  })();
+            const [nodeStatus] = await getNodesStatus({ dockerIndexes: [dockerIndex], onlyActive: false });
+            if (!nodeStatus) console.error(new Error(`Node ${dockerIndex} not found in the monitor`));
+            if (nodeStatus.status === "ONLINE") break;
+            await sleep(5);
+          }
+        })(),
+        sleep(60 * 5).finally(() => {
+          timedout = true;
+          console.error(new Error(`Node ${dockerIndex} did not come online in 5 minutes.`));
+        }),
+      ]);
+    })();
+    updatingInfo.provenToBeOnline = true;
+  }
 
   // restore backup
-  thisIf: if (backup) {
+  thisIf: if (!updatingInfo.backupRestored && backup) {
     // wait until a new block dir is created
     const doesExist = await (async (timedout = false, doesExist = false) => {
       await Promise.race([
@@ -166,10 +196,13 @@ async function restoreAndExit(
 
     // start the docker
     await docker(`inc_mainnet_${node.dockerIndex}`, "start");
+
+    updatingInfo.backupRestored = true;
   }
 
   // stop the docker if it was stopped before
-  if (!info.docker.running) await docker(`inc_mainnet_${node.dockerIndex}`, "stop");
+  if (!info.docker.running || !updatingInfo.dockerWasOnline)
+    await docker(`inc_mainnet_${node.dockerIndex}`, "stop");
 
   await changeNode({ _id: node._id }, { $set: { inactive: false } });
   addNodeToConfigs(node.dockerIndex, node.name, node.validatorPublic);
@@ -178,7 +211,6 @@ async function restoreAndExit(
   await updateTagInDB(node);
 
   console.log(
-    `Node ${node.dockerIndex} updated from ${info.docker.tag} ` +
-      `to ${await getLatestTag()}${backup ? " with backup" : ""}`
+    `Node ${node.dockerIndex} updated from ${info.docker.tag} ` + `to ${latestTag}${backup ? " with backup" : ""}`
   );
 }
